Fixed bugs + added 'Add Menu Item' functionality
- Fixed the search so that:

If the restaurant exists in the restaurant database, navigate to the full menu, this page also has a button to add menu items.

If the restaurant exists in the user's saved menus, navigate to the saved menu. This page has an option only remove an item or edit the notes of the menu item

If the restaurant exists in the user's created menus, navigate to the created menu. This page has an option to edit  all fields of an item and to delete the menu item.

The user can also now add a menu item, either from the restaurant's full menu or to their created menus.

Bug found: Search function: when searching, if the menu is not either in the savedMenus or cretedMenus of the user and also not in the restaurant database, then the user needs to be brought to a page where they can create the menu - or an alert giving them the option to create a menu or go back to the search results.