Finished recommendation service
Users are now given recommendations from restaurants - the menu items that the user has is compared to the menu items in the restaurant database and will only display menu items that the user doesnt already have saved.

It also will highlight the user's allergens in red so the user can see that that menu item has the allergy in it, but can still have it.

All menus will have a banner image.
Created menus will have the location image from google
Saved menus will have the location image from google based on the preferred location
Full menus will have the logo from the restaurant database.

When a user searches for a menu, and it's not on the application, from the search results, the user will have a modal appear giving them the option to create a menu and when they do , the menu location data is passed through and prepopulates the address and the city based off of the search result.

Changed the icons in the tab bar.