Adjusted recommendations page
Previously, it would not filter out menu items that the user has already saved. Now when the user clicks 'Add to Menu', the item disappears from the page. Same with the modal. Also, pulled in the thumbnails for the restaurants so they will sit in the Accordion Item similar to the layout of the menu items.